+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added /compact command to compact conversation history
 - prompts starting with '/' are checked if they are path like before checking them as commands
+- properly closing deleted MCP servers
 
 ## [0.17.0]